+++ packages/components/data-table-manager/src/data-table-settings/droppable-panel/messages.js
@@ -1,0 +1,11 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  loadMoreAttributesHint: {
+    id:
+      'UIKit.DataTableManager.ColumnManager.DroppablePanel.loadMoreAttributesHint',
+    description:
+      'Indicate that the user can use the search input to load more attributes',
+    defaultMessage: 'Load more attributes by using the search input above',
+  },
+});