+++ packages/components/buttons/primary-button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @commercetools-uikit/primary-button
 
+## 10.27.1
+
+### Patch Changes
+
+- Updated dependencies [[`0c09ae4`](https://github.com/commercetools/ui-kit/commit/0c09ae4c3ceaeb65af57bdfcb7ed07c64eff8cb4)]:
+  - @commercetools-uikit/spacings-inline@10.27.1
+
 ## 10.27.0
 
 ### Patch Changes